requirements.txt
fitz==0.0.1.dev2
nltk==3.8.1
openai==1.23.6
pydub==0.25.1
python-dotenv==1.0.1
